requirements.txt
kafka-python
streamlit
pandas
numpy
watchdog
jupyter